+++ app/controllers/template_options_controller.rb
@@ -24,16 +24,15 @@ class TemplateOptionsController < ApplicationController
     return unless (org.present? && !org.new_record?) || (funder.present? && !funder.new_record?)
 
     if org.present? && !org.new_record?
-      # Load the funder's template(s) minus the default template (that gets swapped
-      # in below if NO other templates are available)
-      @templates = Template.latest_customizable.where(org_id: funder.id, is_default: false).sort_by(&:title).to_a
-      # Swap out any organisational cusotmizations of a funder template
+      # Load the funder's template(s)
+      # (`Template.default` belongs to `funder` and will be fetched as part of this query)
+      @templates = Template.latest_customizable.where(org_id: funder.id).sort_by(&:title).to_a
+      # Wherever possible, replace funder templates with organisational customizations
       @templates = @templates.map do |tmplt|
         customization = Template.published
                                 .latest_customized_version(tmplt.family_id,
                                                            org.id).first
-        # Only provide the customized version if its still up to date with the
-        # funder template!
+        # Only provide the customized version if it's still up to date with the funder template
         if customization.present? && !customization.upgrade_customization?
           customization
         else
@@ -42,34 +41,19 @@ class TemplateOptionsController < ApplicationController
       end
       # We are using a default funder to provide with the default templates, but
       # We still want to provide the organization templates.
-      # If the no funder was specified OR the funder matches the org
-      # if funder.blank? || funder.id == org&.id
       # Retrieve the Org's templates
       @templates << Template.published.organisationally_visible.where(org_id: org.id,
                                                                       customization_of: nil).sort_by(&:title).to_a
     else
       # if'No Primary Research Institution' checkbox is checked,
       # only show publicly available template without customization
-      @templates << Template.default if Template.default.present?
-      @templates << Template.published.publicly_visible.where(org_id: funder.id, customization_of: nil,
-                                                              is_default: false).sort_by(&:title).to_a
+      # (`Template.default` belongs to `funder` and will be fetched as part of this query)
+      @templates << Template.published.publicly_visible.where(org_id: funder.id,
+                                                              customization_of: nil).sort_by(&:title).to_a
     end
     @templates = @templates.flatten
-    # DMP Assistant: We do not want to include not customized templates from default funder
-    # Include customizable funder templates
-    # @templates << funder_templates = Template.latest_customizable
-    # Always use the default template
-    if Template.default.present? && org.present?
-      # Fetch the org’s latest customization of the default template
-      customization = Template.published.latest_customized_version(Template.default.family_id, org.id).first
-      # In short: check if `customization` is based on `Template.default`
-      # If so, use `customization`; otherwise, fall back to `Template.default`
-      customization = Template.default unless customization.present? && !customization.upgrade_customization?
-      @templates.select! { |t| t.id != Template.default.id && t.id != customization.id }
-      # We want the default template to appear at the beggining of the list
-      @templates.unshift(customization)
-    end
-    @templates = @templates.uniq
+
+    @templates = sort_templates(@templates, org)
   end
   # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
   # rubocop:enable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
@@ -83,5 +67,41 @@ class TemplateOptionsController < ApplicationController
 
   def org_params
     %i[id name url language abbreviation ror fundref weight score]
+  end
+
+  # Orders templates for dropdown:
+  # 1. Org templates
+  # 2. Priority funder templates (Simplified, Default)
+  # 3. Remaining funder templates
+  def sort_templates(templates, org)
+    # Get the funder templates with the Priority templates ordered in front
+    sorted_funder_templates = sort_funder_templates(templates)
+
+    # If an org was selected and the selected org is not the default funder
+    if org&.id && org.id != Rails.application.config.default_funder_id
+      # Return the templates with the org templates ordered in front
+      org_templates = templates.select { |t| t.org_id == org.id }
+      org_templates + sorted_funder_templates
+    # else, either no org or default_funder was selected
+    else
+      # All of the templates are funder templates
+      sorted_funder_templates
+    end
+  end
+
+  # Orders funder templates:
+  # 1. Priority (Simplified, Default)
+  # 2. Remaining funder templates
+  def sort_funder_templates(templates)
+    funder_templates = templates.select { |t| t.org_id == Rails.application.config.default_funder_id }
+
+    # Identify priority templates
+    # NOTE: This will have to be updated if the `simplified` template is ever renamed
+    simplified = funder_templates.find { |t| t.title.start_with?('Alliance Simplified Template') }
+    default   = funder_templates.find(&:is_default)
+    priority  = [simplified, default].compact
+
+    # Return priority first, then remaining funder templates
+    priority + (funder_templates - priority)
   end
 end